requirements.txt
langchain-core
langchain-huggingface
pydantic
openai
pinecone-client
langchain
langchain-community
python-dotenv
langchain-pinecone
jinja2
langchain-ollama
google-api-core
googleapis-common-protos
protobuf
grpcio
grpcio-tools